Two more shifts into the sign bit, and one of them is three lines from its own fix
  lib/msun/src/e_fmod.c:113   lx = (lx>>n)|((u_int32_t)hx<<(32-n));
  lib/msun/src/e_fmod.c:116   lx = (hx<<(32-n))|(lx>>n); hx = sx;

Same file, same function, three lines apart. Line 113 casts and line 116
does not. hx is int32_t (line 29) and at that point holds |x|'s high word
with the implicit bit set, so up to 0x001FFFFF; n is in [21,31] there, so
the shift is up to 11 and 0x1FFFFF << 11 is 0xFFFFF800 - past INT_MAX.
C17 6.5.7p4.

  lib/msun/src/s_nexttoward.c:46
      INSERT_WORDS(x,uy.bits.sign<<31,1);

uy.bits.sign is `unsigned int :1`. A one-bit unsigned bitfield promotes to
INT, because int represents 0 and 1 - so this is 1 << 31 on a signed int,
which is the s_rint.c finding again in a different disguise. Reached on
every negative argument with x == 0.0.

  $ gcc -fsanitize=undefined -fno-sanitize-recover=all
  runtime error: left shift of 1 by 31 places cannot be represented
                 in type 'int'

That is the third time this shape has turned up with its own correct form
already present nearby: stdc_first_leading_zero.c had the cast the other
three stdbit files omitted, and now e_fmod.c:113 has the one e_fmod.c:116
omits.

Bit-preserving: 3,300,002 cases across every n in [21,31] and both sign
values, zero mismatches, under UBSan.

CBMC after: nexttoward VERIFICATION SUCCESSFUL. fmod still reports three
failures and all three are UNWINDING ASSERTIONS at --unwind 8 - loops that
want more iterations, which is BOUNDED and not a defect. The signed shift
is gone.

Seven UB defects fixed now, all in code reached by ordinary arguments.

+++ hbsd/src/lib/msun/src/e_fmod.c
@@ -113,7 +113,7 @@ fmod(double x, double y)
 		lx = (lx>>n)|((u_int32_t)hx<<(32-n));
 		hx >>= n;
 	    } else if (n<=31) {
-		lx = (hx<<(32-n))|(lx>>n); hx = sx;
+		lx = ((u_int32_t)hx<<(32-n))|(lx>>n); hx = sx;
 	    } else {
 		lx = hx>>(n-32); hx = sx;
 	    }
